refactor: change the folder location
